requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiohttp==3.10.5
redis==5.0.8
asyncpg==0.29.0
SQLAlchemy==2.0.0
sqlalchemy-repr==0.0.2